Add Setup.local and specify which modules to link statically
This fixes build problems on OS X

+++ build-python.sh
@@ -15,15 +15,20 @@ else
 fi
 
 pushd "${PYTHON_DIR}"
-./configure
-make ${PYTHON_BINARY} Parser/pgen
-mv ${PYTHON_BINARY} hostpython
-mv Parser/pgen Parser/hostpgen
-make distclean
+
+if [ ! -f "${PYTHON_BINARY}" ]
+then
+	./configure
+	make ${PYTHON_BINARY} Parser/pgen
+	mv ${PYTHON_BINARY} hostpython
+	mv Parser/pgen Parser/hostpgen
+	make distclean
+fi
 
 CONFIG_SITE=./config.site emconfigure ./configure --without-threads --without-pymalloc --disable-shared --disable-ipv6 --without-gcc --host=asmjs-unknown-emscripten --build=${BUILD_TRIPPLE} --prefix=${EMSCRIPTEN_USR_DIR}
+cp ../Setup.local Modules/
 emmake make HOSTPYTHON=./hostpython HOSTPGEN=./Parser/hostpgen CROSS_COMPILE=yes
-emmake make install
+emmake make install HOSTPYTHON=./hostpython HOSTPGEN=./Parser/hostpgen CROSS_COMPILE=yes
 make distclean
 popd
